Change map back to storing FLOAT

+++ SEG/src/explore/ExploreTest.java
@@ -17,7 +17,7 @@ public class ExploreTest {
 			if (path != null) {
 				int size = path.size();
 				Point end = path.get(size - 1);
-				byte found = Math.random() < 0.4 ? Map.EMPTY : Map.OCCUPIED;
+				float found = Math.random() < 0.4 ? Map.EMPTY : Map.OCCUPIED;
 				map.setValue(end.x, end.y, found);
 				
 				if (found < 0.4) //Target cell is EMPTY, start from it.
